calculate RPM closes #11
